requirements.txt


openai-whisper
flask
moviepy
pytube3
setuptools-rust